% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cde_get_data.R
\name{cde_get_data}
\alias{cde_get_data}
\title{Get Arrests Data}
\usage{
cde_get_data(
  offense,
  state,
  gender = "male",
  since = "1991",
  until = "2010",
  range
)
}
\arguments{
\item{offense}{The type of crime}

\item{state}{The state of the offenders (eg. CA)}

\item{gender}{The gender of the offenders (default male)}

\item{since}{Beginning year of the offense (default 1991)}

\item{until}{Ending year of the offense (default 2010)}

\item{range}{Age range of the offenders (check CDE classifications for range)}
}
\value{
a data frame containing the arrest counts
}
\description{
Get arrests data from Crime Data Explorer (CDE)
}
\examples{
cde_get_data(offense="burglary", state="CA", range="range_13_14")
}
