% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_get_series.R
\name{fred_get_series}
\alias{fred_get_series}
\title{Fetch Series from FRED}
\usage{
fred_get_series(series_id, observation_start, observation_end)
}
\arguments{
\item{series_id}{The id of the series}

\item{observation_start}{Starting time of the fetched series}

\item{observation_end}{Ending time of the fetched series}
}
\value{
a data frame containing the series
}
\description{
Fetch a series from FRED and return as a data frame
}
\examples{
fred_get_series(series_id="BOPBCA", observation_start="1960-01-01", observation_end="2010-12-31")
}
