% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_fold_CV.R
\name{k_fold_CV}
\alias{k_fold_CV}
\title{K-Fold Cross Validation}
\usage{
k_fold_CV(y, x, K = 5)
}
\arguments{
\item{y}{The dependent variable}

\item{x}{The explanatory variable}

\item{K}{The number of folds to use (default 5)}
}
\value{
R-squared and RMSE of the cross-validation
}
\description{
Perform K-fold cross-validation for a simple linear regression with one explanatory
variable
}
\examples{
k_fold_CV(y, x, K=10)
}
