% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagrange.R
\name{lagrange}
\alias{lagrange}
\title{Lagrange-Form Interpolation Polynomial Estimation}
\usage{
lagrange(m, x, y, z)
}
\arguments{
\item{m}{Number of data points given}

\item{x}{Vector that contains the x-coordinates of the data points}

\item{y}{Vector that contains the y-coordinates of the data points}

\item{z}{Point of query whose y-value is the interest}
}
\value{
The y-value of the point of query, estimated by the interpolation polynomial
}
\description{
Use the Lagrange form to estimate the interpolation polynomial for the given data
points. This function then returns the y-value for a user-inputted point of query
}
\examples{
lagrange(m, x, y, z)
}
