% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cointegration.R
\name{test_cointegration}
\alias{test_cointegration}
\title{Test Cointegration}
\usage{
test_cointegration(vec1, vec2, percent = 0.05)
}
\arguments{
\item{vec1}{The first series}

\item{vec2}{The second series}

\item{percent}{The significance level (default 0.05)}
}
\value{
1 if the two series are cointegrated; 0 otherwise
}
\description{
Test if two series of data are cointegrated at a specific significance level
}
\examples{
test_cointegration(vec1, vec2, percent=0.01)
}
